requirements.txt
jupyter~=1.0.0
pip~=23.3.1
pathlib~=1.0.1
panel~=1.3.8
attrs~=23.1.0
zlib~=1.2.13
distro~=1.8.0
wheel~=0.41.2
openssl~=3.0.13
pillow~=10.2.0
docutils~=0.18.1
sphinx~=5.0.2
tornado~=6.3.3
jinja2~=3.1.3
setuptools~=68.2.2
packaging~=23.2
pyparsing~=3.0.9
zipp~=3.17.0
typing_extensions~=4.9.0
future~=0.18.3
langchain
langchain-openai
langchain-core
python-dotenv
langserve
fastapi
uvicorn
pymongo
dnspython
pymongo[srv]
openai
requests
pytesseract
langchain-openai
pillow
tesseract-ocr
opencv-python
streamlit
